cleanup - move common logic to super class
